Move clear changes to own component
